+++ src/zustand/modalState.js
@@ -1,0 +1,12 @@
+import { create } from "zustand";
+
+const useModalState = create((set, get) => ({
+  modalIsOpen: false,
+
+  handleModal: () => {
+    const newState = { modalIsOpen: !get().modalIsOpen };
+    set(newState);
+  },
+}));
+
+export default useModalState;
